fix entity damaged issue and onload
The mod was found to not register any events on loading with an existing map of the current mod version. this was caused by a change in 0.1.1 to fix an error with OnLoad. The solution is you need OnLoad to reRegister the events for when the mod version is the same. If the mod is being upgraded the mod settings may not be present yet as OnLoad is called before OnConfigurationChange. In this case OnLoad does nothing event wise and OnConfigurationChnage then fixes everything.
The Entity Damaged event error has been fixed by checking of type player and then seperately checking if the Character entity has a player.